requirements.txt
dgl>=1.1.2
easydict>=1.9
numpy>=1.23
ogb>=1.3.6
scikit-learn>=1.1.3
torch>=2.0.0
tqdm
pyyaml
